requirements.txt
streamlit
requests
qrcode
Pillow
fpdf
pdfplumber
pytesseract

